修复BUG

+++ zhaoyunpeng/Assets/_Scripts/Main/Backpack/backpack_manger.cs
@@ -64,7 +64,7 @@ public class backpack_manger : MonoBehaviour {
     //模拟数据库取物体
 	private void load(){
         ItemList = new Dictionary<int, item>();
-        Weapon w1 = new Weapon(0,"枪","伤害力强大","");
+        Weapon w1 = new Weapon(0,"M4A1","伤害力普通","");
         Weapon w2 = new Weapon(1, "匕首", "伤害力一般","");
         Weapon w3 = new Weapon(2, "木棒", "伤害力普通","");
         Weapon w4 = new Weapon(3, "板砖", "伤害力弱","");
@@ -105,8 +105,9 @@ public class backpack_manger : MonoBehaviour {
         }
         else{
             item item = Itemmodel.GetItem(gridTransform.name);
-            
-            DragItemUI.updateItem(item.Name);
+
+            //Debug.Log("123");
+            DragItemUI.updateItem(item.Id);
             Destroy(gridTransform.GetChild(0).gameObject);
             
             isDrag = true;
@@ -114,42 +115,45 @@ public class backpack_manger : MonoBehaviour {
     }
     private void GridUI_OnLeftEndDrag(Transform prevtransform, Transform eventtransform)
     {
-
-        isDrag = false;
-        DragItemUI.hidden();
-        if (eventtransform == null)//扔东西
+        if (isDrag)
         {
-            item item = Itemmodel.GetItem(prevtransform.name);
+            isDrag = false;
+            DragItemUI.hidden();
+            if (eventtransform == null)//扔东西
+            {
+                item item = Itemmodel.GetItem(prevtransform.name);
             
-            this.CreatNewItemPlane(item);
-            Itemmodel.DeleteItem(prevtransform.name);
+                this.CreatNewItemPlane(item);
+                Itemmodel.DeleteItem(prevtransform.name);
             
 
-        }
-        else if (eventtransform.tag == "Grid") {//拖到另一个格子或者当前格子
-                 if (eventtransform.childCount == 0)//直接扔进去
-                 {
-                     item item = Itemmodel.GetItem(prevtransform.name);
-                     this.CreatNewItem(item, eventtransform);
-                     Itemmodel.DeleteItem(prevtransform.name);
+            }
+            else if (eventtransform.tag == "Grid") {//拖到另一个格子或者当前格子
+                     if (eventtransform.childCount == 0)//直接扔进去
+                     {
+                         item item = Itemmodel.GetItem(prevtransform.name);
+                         this.CreatNewItem(item, eventtransform);
+                         Itemmodel.DeleteItem(prevtransform.name);
                      
-                 }
-                 else //交换
-                 {
-                     Destroy(eventtransform.GetChild(0).gameObject);
-                     item prevGridItem = Itemmodel.GetItem(prevtransform.name);
-                     item enterGridItem = Itemmodel.GetItem(eventtransform.name);
-                     this.CreatNewItem(prevGridItem, eventtransform);
-                     this.CreatNewItem(enterGridItem, prevtransform);
-                 }
+                     }
+                     else //交换
+                     {
+                         Destroy(eventtransform.GetChild(0).gameObject);
+                         item prevGridItem = Itemmodel.GetItem(prevtransform.name);
+                         item enterGridItem = Itemmodel.GetItem(eventtransform.name);
+                         this.CreatNewItem(prevGridItem, eventtransform);
+                         this.CreatNewItem(enterGridItem, prevtransform);
+                     }
         
         
+            }
+            else//拖到UI其他地方
+            {
+                item item = Itemmodel.GetItem(prevtransform.name);
+                this.CreatNewItem(item, prevtransform);
+            }
         }
-        else//拖到UI其他地方
-        {
-            item item = Itemmodel.GetItem(prevtransform.name);
-            this.CreatNewItem(item, prevtransform);
-        }
+        
         
 
     }
@@ -163,14 +167,13 @@ public class backpack_manger : MonoBehaviour {
         GameObject Parent = GameObject.FindGameObjectWithTag("boot");
         itemGo.transform.parent = Parent.transform;
         itemGo.transform.position = camera.transform.TransformPoint(0,0,1);
-        itemGo.transform.localScale = new Vector3(0.2f, 0.2f, 0.2f);
-        
     }
 
     private void CreatNewItem(item item,Transform parent)
     {
-        GameObject Itemprefabs = Resources.Load<GameObject>("_Prefabs/Item");
-        Itemprefabs.GetComponent<ItemUI>().updateItem(item.Name);
+        string src = "_Prefabs" + "/" + "Item";
+        GameObject Itemprefabs = Resources.Load<GameObject>(src);
+        Itemprefabs.GetComponent<ItemUI>().updateItem(item.Id);
         GameObject itemGo = GameObject.Instantiate(Itemprefabs);
         itemGo.transform.SetParent(parent);
         itemGo.transform.localPosition = Vector3.zero;
